Added spinners

+++ Cheese App/Views/MyCheesesView.swift
@@ -7,6 +7,11 @@
 
 import SwiftUI
 
+class MyCheesesViewModel: ObservableObject {
+    @Published var isLoading: Bool = true
+    
+}
+
 struct MyCheesesView: View {
     @State public var showCupboardPopover = false
     @State public var showCheesePopover = false
@@ -16,6 +21,7 @@ struct MyCheesesView: View {
     @AppStorage("profileId") var profileId: String?
     @State private var cupboards: [Cupboard]?
     @Binding var selectedTab: Tab
+    @StateObject var viewModel = MyCheesesViewModel()
     
     var body: some View {
         NavigationStack {
@@ -32,52 +38,57 @@ struct MyCheesesView: View {
                         }
                         .font(.custom("IowanOldStyle-Roman", size: 24))
                         
-                        
-                        
-                        List{
-                            Section{
-                                ForEach(cupboards ?? []) { cupboard in
-                                    if cupboard.name != "Created By Me" {
-                                        NavigationLink(destination: CupboardListView(cupboardId: cupboard.id!, selectedTab: $selectedTab, showAddCheeseButton: true)) {
-                                            Text(cupboard.name ?? "")
-                                                .font(.custom("IowanOldStyle-Roman", size: 24))
-                                                .frame(maxWidth: .infinity, alignment: .leading)
-                                                .fontWeight(.bold)
+                        if(viewModel.isLoading){
+                            VStack{
+                                ProgressView() // Spinner shown when loading
+                                    .progressViewStyle(CircularProgressViewStyle())
+                                    .scaleEffect(1.5) // Make the spinner larger if needed
+                            }.frame(maxWidth: .infinity, maxHeight: .infinity).background(CustomColors.background)
+                        } else {
+                            List{
+                                Section{
+                                    ForEach(cupboards ?? []) { cupboard in
+                                        if cupboard.name != "Created By Me" {
+                                            NavigationLink(destination: CupboardListView(cupboardId: cupboard.id!, selectedTab: $selectedTab, showAddCheeseButton: true)) {
+                                                Text(cupboard.name ?? "")
+                                                    .font(.custom("IowanOldStyle-Roman", size: 24))
+                                                    .frame(maxWidth: .infinity, alignment: .leading)
+                                                    .fontWeight(.bold)
+                                            }
+                                            .padding()
+                                        } else {
+                                            NavigationLink(destination: CupboardListView(cupboardId: cupboard.id!, selectedTab: $selectedTab, showAddCheeseButton: false)) {
+                                                Text(cupboard.name ?? "")
+                                                    .font(.custom("IowanOldStyle-Roman", size: 24))
+                                                    .frame(maxWidth: .infinity, alignment: .leading)
+                                                    .fontWeight(.bold)
+                                            }
+                                            .deleteDisabled(true)
+                                            .padding()
                                         }
-                                        .padding()
-                                    } else {
-                                        NavigationLink(destination: CupboardListView(cupboardId: cupboard.id!, selectedTab: $selectedTab, showAddCheeseButton: false)) {
-                                            Text(cupboard.name ?? "")
-                                                .font(.custom("IowanOldStyle-Roman", size: 24))
-                                                .frame(maxWidth: .infinity, alignment: .leading)
-                                                .fontWeight(.bold)
-                                        }
-                                        .padding()
                                     }
+                                    .onDelete(perform: { indexSet in
+                                        let idsToDelete = indexSet.map { cupboards![$0].id }
+                                        if (idsToDelete.count == 0){
+                                            return
+                                        }
+                                        Task{
+                                            await Database().deleteCupboard(cupboardId: idsToDelete[0]!)
+                                            
+                                            
+                                        }
+                                        print("deleted")
+                                    })
+                                    
+                                    .listRowBackground(CustomColors.background)
+                                    
+                                    
                                 }
                                 
-                                .onDelete(perform: { indexSet in
-                                    let idsToDelete = indexSet.map { cupboards![$0].id }
-                                    if (idsToDelete.count == 0){
-                                        return
-                                    }
-                                    Task{
-                                        await Database().deleteCupboard(cupboardId: idsToDelete[0]!)
-                                        
-                                        
-                                    }
-                                    print("deleted")
-                                })
-                                
-                                .listRowBackground(CustomColors.background)
-                                
-                                
                             }
-                            
+                            .scrollContentBackground(.hidden)
+                            .listStyle(PlainListStyle())
                         }
-                        .scrollContentBackground(.hidden)
-                        .listStyle(PlainListStyle())
-                        
                         Spacer()
                         HStack{
                             Button(action: {
@@ -126,6 +137,7 @@ struct MyCheesesView: View {
                     cupboards = await Database().getUserCupboards(profileId: profileId!)
                     
                 }
+                viewModel.isLoading = false
             }
             
         }
@@ -134,6 +146,13 @@ struct MyCheesesView: View {
     
 }
 
-//#Preview {
-//    MyCheesesView()
-//}
+struct MyCheesesViewPreview: View {
+    @State var selectedTab = Tab.home
+    var body: some View {
+        MyCheesesView(selectedTab: $selectedTab)
+    }
+}
+
+#Preview {
+    MyCheesesViewPreview()
+}
